Introduced r2s_debug to keep fispact workplaces.

+++ scripts/collapse_s2.sh
@@ -5,8 +5,8 @@
 wdr="$r2s_scratch"/r2s_c/$1/$2/$3           # Collapse working folder (see collapse_s1.sh)
 flx="$r2s_scratch"/r2s_w/fluxes.$1.$2.$3    # File with flux spectrum
 
-rm -rf "$wdr"
-rm -rf "$flx"
+[  -z "$r2s_debug"  ] && rm -rf "$wdr"
+[  -z "$r2s_debug"  ] && rm -rf "$flx"
 
 mv "$r2s_scratch"/r2s_w/cgi.$1.$2.$3 "$r2s_out"
